Create a ResourceLocatorBase
